[ansatz] Reorganized the pulses module.
This is part of a cleanup and testing process.
