+++ pulse_opt/pulses/__init__.py
@@ -1,6 +1,6 @@
 from . import linear_pulses
-from . import gaussian_pulses
-from . import power_pulses
-from . import fourier_pulses
+from . import gaussian_factory
+from . import power_factory
+from . import fourier_factory
 from . import utilities
 from . import visualizations
