+++ tests/pulses/test_fourier_pulses.py
@@ -1,10 +1,2 @@
 import pytest
 import numpy as np
-
-from pulse_opt.pulses.fourier_pulses import FourierPulse
-
-
-def test_fourier_pulse_init():
-    FourierPulse()
-
-
