+++ tests/pulses/test_gaussian_pulses.py
@@ -1,12 +1,2 @@
 import pytest
 import numpy as np
-
-from pulse_opt.pulses.gaussian_pulses import (
-    GaussianSuperpositionPulse,
-)
-
-
-def test_gaussian_superposition_pulse_init():
-    GaussianSuperpositionPulse()
-
-
